A lot of changes to the joystick handling and the animations system

+++ game/engine/2d_sprites/include/Sprite.h
@@ -39,7 +39,7 @@ private:
     Animation *_animation;
     std::string _animation_name;
     Uint8 _animation_frame_index;
-    Uint8 _animation_frame_delay;
+    float _animation_frame_delay;
     bool _animating;
 
     void setAnimationFrame(Uint8 frame_index);
